+++ app/Http/Controllers/CategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Category;
+use Illuminate\Support\Facades\DB;
 
 class CategoryController extends Controller
 {
@@ -83,7 +84,7 @@ class CategoryController extends Controller
         $category = Category::find($id);
         $category->title = $request->input('category-name');
         $category->save();
-        return redirect()->route('category.list');
+        return redirect()->route('category.list')->with("success","Sửa thành công");
     }
 
     /**
@@ -94,9 +95,17 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
-        $category = Category::find($id);
-        $category->delete();
-        return redirect()->route('category.list');
+        $check = DB::select('select * from products where category_id = :id', ['id' => $id]);
+        if (is_array($check)) {
+            if (count($check) > 0) {
+                return back()->with("invalid", "Hiện có một số sản phẩm đang có danh mục này");
+            }
+            else{
+                $category = Category::find($id);
+                $category->delete();
+                return redirect()->route('category.list')->with("success","Xóa thành công");
+            }
+        }
     }
 
     /**
@@ -110,7 +119,7 @@ class CategoryController extends Controller
         $category = Category::find($id);
         $category->status = 0;
         $category->save();
-        return redirect()->route('category.list');
+        return redirect()->route('category.list')->with("success","Vô hiệu hóa thành công");
     }
 
     /**
@@ -124,6 +133,6 @@ class CategoryController extends Controller
         $category = Category::find($id);
         $category->status = 1;
         $category->save();
-        return redirect()->route('category.list');
+        return redirect()->route('category.list')->with("success","Mở thành công");
     }
 }
